Removed obsolete `readPed` and moved `makeA` back to pedigree
